moved TreatmentPlan entity to the Patients aggregate
