+++ DigiDent.Domain/ClinicCoreContext/Patients/Errors/TreatmentPlanErrors.cs
@@ -1,7 +1,7 @@
-﻿using DigiDent.Domain.ClinicCoreContext.Visits.ValueObjects;
+﻿using DigiDent.Domain.ClinicCoreContext.Patients.ValueObjects;
 using DigiDent.Domain.SharedKernel.ReturnTypes;
 
-namespace DigiDent.Domain.ClinicCoreContext.Visits.Errors;
+namespace DigiDent.Domain.ClinicCoreContext.Patients.Errors;
 
 public static class TreatmentPlanErrors
 {
