+++ DigiDent.Domain/ClinicCoreContext/Patients/ValueObjects/TreatmentPlanId.cs
@@ -1,5 +1,5 @@
 ﻿using DigiDent.Domain.SharedKernel.Abstractions;
 
-namespace DigiDent.Domain.ClinicCoreContext.Visits.ValueObjects.Ids;
+namespace DigiDent.Domain.ClinicCoreContext.Patients.ValueObjects;
 
 public record TreatmentPlanId(Guid Value): ITypedId<Guid>;
