+++ DigiDent.Domain/ClinicCoreContext/Shared/Errors/TimeErrors.cs
@@ -1,12 +1,12 @@
 ﻿using DigiDent.Domain.ClinicCoreContext.Shared.ValueObjects;
 using DigiDent.Domain.SharedKernel.ReturnTypes;
 
-namespace DigiDent.Domain.ClinicCoreContext.Visits.Errors;
+namespace DigiDent.Domain.ClinicCoreContext.Shared.Errors;
 
 public static class TimeErrors
 {
     public static Error DurationIsNotPositive
-        => new Error(
+        => new (
             ErrorType.Validation,
             nameof(TimeDuration),
             "The duration must be positive.");
